Rework install process for web apps and plugins
This is the first of a few commits that will reinvent this process
to make it easier to install third-party plugins into girder,
including pip-installed girder instances.
